+++ CHANGELOG.rst
@@ -19,7 +19,7 @@ file carefully before upgrading.
   the specified migration to be applied, plus any dependencies. In the case
   of rollback, this removes the specified revision and any other migrations
   that depend upon it.
-* Added 'mark' command to allow migrations to be marked in the
+* Added 'mark' and 'unmark' commands to allow migrations to be marked in the
   database without actually running them
 * Transaction handling has changed. Each migration now always runs in a
   single transaction, with individual steps running in nested transactions